+++ js/components/flipclock.js
@@ -12,6 +12,8 @@ var DT_flipclock = {
     return {
       width: 12,
       scale: 1,
+      showSecoonds: !settings['hide_seconds'],
+      clockFace: settings['shorttime'].match(/A/i) ? 12 : 24,
     };
   },
   run: function (me) {
@@ -22,10 +24,9 @@ var DT_flipclock = {
     );
 
     FlipClock($(me.mountPoint + ' .dt_content'), {
-      clockFace: settings['shorttime'].match(/A/i)
-        ? 'TwelveHourClock'
-        : 'TwentyFourHourClock',
-      showSeconds: !settings['hide_seconds'],
+      clockFace:
+        me.block.clockFace == 12 ? 'TwelveHourClock' : 'TwentyFourHourClock',
+      showSeconds: me.block.showSeconds,
     });
   },
 };