+++ docs/releasenotes/releasenotes.rst
@@ -19,6 +19,7 @@ Enhancements
 * Calendar: New block parameter ``emptytext`` to define the text to show where there are no calendar appointments. Only works for the new calendar block. See :ref:`newcalendar`
 * Custom graph: aggregate parameter can be an array to specify different aggregation methods per data element. See :ref:`groupBy`
 * Support for device (sub)type Managed Counter
+* Flipclock: New block parameters showSeconds (true or false) and clockFace (12 or 24)
 
 Fixes
 ~~~~~~